// fe/src/lib/api/generated/model/routineConcernValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type RoutineConcernValue = typeof RoutineConcernValue[keyof typeof RoutineConcernValue];


export const RoutineConcernValue = {
  FAMILY: 'FAMILY',
  FINANCIAL: 'FINANCIAL',
  ACADEMIC: 'ACADEMIC',
  FRIENDS: 'FRIENDS',
  CLASSMATES: 'CLASSMATES',
  VICES: 'VICES',
  LOVE_LIFE: 'LOVE_LIFE',
  SLEEPING_PROBLEMS: 'SLEEPING_PROBLEMS',
  SUICIDAL_THOUGHT_TENDENCY: 'SUICIDAL_THOUGHT_TENDENCY',
  DORM_BOARDING_HOUSE: 'DORM_BOARDING_HOUSE',
  PAST_PAINFUL_EXPERIENCE: 'PAST_PAINFUL_EXPERIENCE',
  OTHER: 'OTHER',
} as const;