// fe/src/lib/api/generated/model/routineIntakePayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { RoutineConcernValue } from './routineConcernValue';

export interface RoutineIntakePayload {
  academic_goals?: string;
  career_goals?: string;
  college_adjustment_and_peer_group?: string;
  college_experience?: string;
  concerns?: RoutineConcernValue[];
  concerns_explanation?: string;
  coping_remarks?: string;
  coping_with_college_challenges?: string;
  difficulties_encountered?: string;
  family_description?: string;
  other_concern_specification?: string;
  reason_for_choosing_institution?: string;
  stress_anxiety_causes?: string;
  stress_anxiety_management?: string;
}